requirements.txt
django
djangorestframework
djangorestframework_simplejwt
psycopg2-binary
dj-config-url
python-decouple
drf-yasg
django-cors-headers
pillow
django-filter
celery[redis]
django-jazzmin
